First commit and new MacOS docs section
Created a new mac-specific portion to the docs.
As to why the docs are being created at all, I'm realizing that between my personal life and work I'm devving on Windows, Linux, and Mac, and it's getting confusing when I expect my system to look a certain way and then it's not, so I'm going to try and centralize all my favorite settings so in the future it's just a matter of importing them.